Ignore undocumented bundled extensions/skins
If an extension or skin is bundled with MediaWiki, Canasta will document it in
the values file but Taqasta will omit it unless the extension (or skin)
requires composer dependencies to be installed. Don't include those bundled
extensions and skins in the list of extensions/skins that are present in
Canasta but not in Taqasta.

WIK-2077